requirements.txt
alembic==1.6.5
aniso8601==9.0.1
async-generator==1.10
attrs==21.4.0
boto3==1.20.42
botocore==1.23.42
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
cryptography==36.0.1
Faker==11.3.0
Flask==2.0.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==2.0.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2021.3
randomuser==1.6
requests==2.27.1
s3transfer==0.5.0
selenium==4.1.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.19
text-unidecode==1.3
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
Werkzeug==2.0.1
wsproto==1.0.0
WTForms==2.3.3
